refactor: #15 회원 엔티티 공통 필드 추가
- 생성일, 수정일 필드 추가를 위한 BaseEntity 상속

- 회원 엔티티 패키지 이동
- JoinRequest entity import 경로 변경
